Modified and renamed class files. front end is up and running
