+++ frontend/src/Components/HomePage.tsx
@@ -8,9 +8,9 @@ export const Home = () => {
 };
 
 export function Title() {
-	return <h1>Doggr</h1>;
+	return <h1>Quiz</h1>;
 }
 
 export function Subtitle() {
-	return <h3>Where your pets find love(tm)</h3>;
+	return <h3>Test your knowledge</h3>;
 }
