+++ frontend/src/Components/Match.tsx
@@ -1,10 +1,10 @@
 
 import { Profile } from "@/Components/Profile.tsx";
-import { ProfileType } from "@/DoggrTypes.ts";
+import { ProfileType } from "@/QuizTypes.ts";
 import { useAuth } from "@/Services/Auth.tsx";
-import { getNextProfileFromServer } from "@/Services/HttpClient.tsx";
-import { MatchService } from "@/Services/MatchService.tsx";
-import { PassService } from "@/Services/PassService.tsx";
+// import { getNextProfileFromServer } from "@/Services/HttpClient.tsx";
+// import { MatchService } from "@/Services/MatchService.tsx";
+// import { PassService } from "@/Services/PassService.tsx";
 import { useContext, useEffect, useState } from "react";
 
 export const Match = () => {
@@ -12,45 +12,46 @@ export const Match = () => {
 
 	const auth = useAuth();
 
-	const fetchProfile = () => {
-		getNextProfileFromServer()
-			.then((response) => setCurrentProfile(response))
-			.catch( (err) => console.log("Error in fetch profile", err));
-	};
+	// const fetchProfile = () => {
+	// 	getNextProfileFromServer()
+	// 		.then((response) => setCurrentProfile(response))
+	// 		.catch( (err) => console.log("Error in fetch profile", err));
+	// };
+	//
+	// useEffect(() => {
+	// 	fetchProfile();
+	// }, []);
+	//
+	// const onLikeButtonClick = () => {
+	// 	MatchService.send(auth.userId, currentProfile.id)
+	// 		.then(fetchProfile)
+	// 		.catch(err => {
+	// 			console.error(err);
+	// 			fetchProfile();
+	// 		});
+	// };
+	//
+	// const onPassButtonClick = () => {
+	// 	PassService.send(auth.userId, currentProfile.id)
+	// 		.then(fetchProfile)
+	// 		.catch(err => {
+	// 			console.error(err);
+	// 			fetchProfile();
+	// 		});
+	// };
 
-	useEffect(() => {
-		fetchProfile();
-	}, []);
-
-	const onLikeButtonClick = () => {
-		MatchService.send(auth.userId, currentProfile.id)
-			.then(fetchProfile)
-			.catch(err => {
-				console.error(err);
-				fetchProfile();
-			});
-	};
-
-	const onPassButtonClick = () => {
-		PassService.send(auth.userId, currentProfile.id)
-			.then(fetchProfile)
-			.catch(err => {
-				console.error(err);
-				fetchProfile();
-			});
-	};
-
-	const profile = (
-		<Profile
-			{...currentProfile}
-			onLikeButtonClick={onLikeButtonClick}
-			onPassButtonClick={onPassButtonClick}
-		/>
-	);
+	// const profile = (
+	// 	<Profile
+	// 		{...currentProfile}
+	// 		onLikeButtonClick={onLikeButtonClick}
+	// 		onPassButtonClick={onPassButtonClick}
+	// 	/>
+	// );
 
 	return (
-		<>
-			{profile}
-		</>
+		// <>
+		// 	{profile}
+		// </>
+		<h1>MAtch page</h1>
 	);
 };
