+++ frontend/src/QuizTypes.ts
