+++ frontend/src/Services/HttpClient.tsx
@@ -1,4 +1,4 @@
-import { ProfileType } from "@/DoggrTypes.ts";
+import { ProfileType } from "@/QuizTypes.ts";
 import axios from "axios";
 
 const serverIP = import.meta.env.API_HOST;
@@ -14,8 +14,8 @@ export const httpClient = axios.create({
 	},
 });
 
-export async function getNextProfileFromServer() {
-	const profile =
-		await httpClient.get<ProfileType>("/profile");
-	return profile.data;
-}
+// export async function getNextProfileFromServer() {
+// 	const profile =
+// 		await httpClient.get<ProfileType>("/profile");
+// 	return profile.data;
+// }
