+++ frontend/src/Services/PassService.tsx
@@ -1,7 +1,7 @@
-import { httpClient } from "@/Services/HttpClient.tsx";
-
-export const PassService = {
-	async send(sender_id: number, receiver_id: number) {
-		return httpClient.post("/pass", {sender_id, passee_id: receiver_id});
-	}
-};
+// import { httpClient } from "@/Services/HttpClient.tsx";
+//
+// export const PassService = {
+// 	async send(sender_id: number, receiver_id: number) {
+// 		return httpClient.post("/pass", {sender_id, passee_id: receiver_id});
+// 	}
+// };
